+++ .gitignore
@@ -1,1 +1,3 @@
+__pycache__
 accounts.json
+sessions
