+++ lib/browser.py
@@ -1,0 +1,21 @@
+
+import os
+
+from selenium import webdriver
+from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.chrome.webdriver import WebDriver
+
+from .constants import DESKTOP_USER_AGENT, MOBILE_USER_AGENT
+
+def browserSetup(sessionName: str, headlessMode: bool = False, isMobile: bool = False, lang: str = 'en') -> WebDriver:
+    options = Options()
+    options.add_argument(
+        f'user-agent={MOBILE_USER_AGENT if isMobile else DESKTOP_USER_AGENT}')
+    options.add_argument(f'lang={lang.split("-")[0]}')
+    if headlessMode:
+        options.add_argument('--headless')
+    options.add_argument('log-level=3')
+    currentPath = os.path.dirname(os.path.realpath(__file__))
+    options.add_argument(
+        f'--user-data-dir={currentPath}/sessions/{"mobile" if isMobile else "desktop"}/{sessionName}')
+    return webdriver.Chrome(options=options)
